requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
django==4.2.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-allauth==0.54.0
filelock==3.18.0
identify==2.6.9
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.1.0
PyJWT==2.9.0
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.3
dj-rest-auth==3.0.0